Refactor: se cambiaron nombres de carpetas para mejor estructura
