+++ src/main/java/registro/samtel/horas/models/contract/IRegistroRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import registro.samtel.horas.models.entities.RegistroEntity;
 
 
-public interface IRegistroUsuario extends JpaRepository<RegistroEntity, Long> {
+public interface IRegistroRepository extends JpaRepository<RegistroEntity, Long> {
 
     //crear metodos para la consulta de la base de datos
