+++ src/main/java/registro/samtel/horas/models/contract/IUsuarioRepository.java
@@ -4,7 +4,7 @@ package registro.samtel.horas.models.contract;
 import org.springframework.data.jpa.repository.JpaRepository;
 import registro.samtel.horas.models.entities.UsuarioEntity;
 
-public interface IUsuarioEntity extends JpaRepository<UsuarioEntity, Long> {
+public interface IUsuarioRepository extends JpaRepository<UsuarioEntity, Long> {
 
 
 }
