+++ src/main/java/registro/samtel/horas/services/contract/IRegistroHorasService.java
@@ -1,4 +1,4 @@
 package registro.samtel.horas.services.contract;
 
-public interface IRegistroUsuarioImpl {
+public interface IRegistroHorasService {
 }
