+++ src/main/java/registro/samtel/horas/services/impl/RegistroHorasServiceImpl.java
@@ -3,5 +3,5 @@ package registro.samtel.horas.services.impl;
 import org.springframework.stereotype.Service;
 
 @Service
-public class RegistroUsuarioImpl {
+public class RegistroHorasServiceImpl {
 }
